// src/app/models/interfaces/dashboard.interface.ts
export interface IProfitLossChartResponse {
  lable?: IChartLabelType;
  chartTitle?: string;
  profitLossActiveYear?: any[],
  profitLossLastYear?: any[],
  legend?: string[]
}
// ...
export enum ChartFilterType {
  ThisMonthToDate,
  ThisQuarterToDate,
  ThisFinancialYearToDate,
  ThisYearToDate,
  LastMonth,
  LastQuater,
  LastFiancialYear,
  LastYear,
  Custom
}

export enum ChartType {
  Revenue,
  Expense,
  ProfitLoss
}

export interface IChartLabelType {
  activeYearLabel?: string;
  lastYearLabel?: string;
}